Parts 4.15-4.23

+++ part4/blogger/controllers/blogs.js
@@ -2,18 +2,33 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find()
+  const blogs = await Blog.find().populate('user')
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  const result = await blog.save()
-  response.status(201).json(result)
+  const body = request.body
+  if (request.user === null) {
+    return response.status(401).json({ error: 'Token or UserId missing or not valid' })
+  }
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    likes: body.likes,
+    url: body.url,
+    user: user.id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user')
   if (blog) {
     response.json(blog)
   } else {
@@ -22,6 +37,14 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (request.user === null) {
+    return response.status(401).json({ error: 'Token or UserId missing or not valid' })
+  }
+  const user = request.user
+  const toDelete = await Blog.findById(request.params.id).populate('user')
+  if (user.id.toString() !== toDelete.user.id.toString()) {
+    response.status(401).end()
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
